Have `workflow_transitions` return collection
This is a quality of life improvement for developers who want to lean on Laravel's collections. I currently wrap the return of this function in a collection and would like the ability to have it function that way out of the box.

I don't believe this will introduce regressions in existing code.